requirements.txt
aiohttp==3.8.3
aiosignal==1.2.0
argcomplete==2.0.0
asgiref==3.5.2
async-generator==1.10
async-timeout==4.0.2
attrs==22.1.0
autopep8==1.7.0
beautifulsoup4==4.11.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
cryptography==38.0.1
defusedxml==0.7.1
dj-database-url==1.0.0
Django==4.1.1
django-allauth==0.51.0
django-bootstrap-v5==1.0.11
django-heroku==0.3.1
exceptiongroup==1.0.0rc9
frozenlist==1.3.1
gunicorn==20.1.0
h11==0.14.0
idna==3.4
multidict==6.0.2
oauthlib==3.2.1
outcome==1.2.0
packaging==21.3
psycopg2==2.9.3
pyasn1==0.4.8
pycodestyle==2.9.1
pycparser==2.21
PyJWT==2.5.0
pyparsing==3.0.9
PySocks==1.7.1
python3-openid==3.2.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
selenium==4.5.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
toml==0.10.2
trio==0.22.0
trio-websocket==0.9.2
types-cryptography==3.3.23
tzdata==2022.2
urllib3==1.26.12
userpath==1.8.0
whitenoise==6.2.0
wsproto==1.2.0
yarl==1.8.1
